Add first type: CUTE_Test
This type is the smallest unit of the framework, it corresponds to a
single unit test. It contains a pointer to the test procedure and the
name  of the test procedure, as a string (it can therefore come in
handy to give explicit names to the tests!)

The type is defined in the file `cute_test.h`, and the CUTE_Test type
is not defined opaquely so that instances can live on the stack and not
on the heap and with that the header has no related source: the
functions that manipulate the type are macros.