Add test case structure
The structure represents a set of semantically-related tests to execute
at once.

It can contain a `setUp` and `tearDown` procedure, designed to prepare
or clean up the execution environment for each test, allowing them to be
actually independent. There can also be a global-scope `initiate` and
`terminate` functions, to set up and clean the test case properly.
All these procedures must not take any parameter.