Add compile-time assert for non-C11 C version
`CUTE_compileTimeAssert` is defined with a hacky syntax if the project
is compiled with a version of C older than the 2011 standard (which
introduced the `_Static_assert` keyword).

If the project is compiled with C11 standard though, the macro is
defined with the standard macro `static_assert`, that overwrites this
keyword.