Change test calling exit() with abort()
The function `exit` indicates normal program termination, and no test
should ever terminate the program, at least not normally!
Moreover it wipes clean the environment (flushes and closes streams,
call `atexit` functions...) and that's absolutely not desired.
The call to `exit` is thus replaced with `abort`, which represents an
abnormal program termination (`SIGABRT` signal) and that is much more
expectable for a test behavior.