Remove dependency to assert.h
The project declares a new function, `CUTE_fail`, which is used in the
macro `CUTE_runTimeAssert` if the condition evaluates to false.
The function internally calls `abort(3)`, which triggers a `SIGABRT`
signal that can be handled -- this will allow to generate a means of
recovering from a failed assert.

The macro `CUTE_compileTimeAssert` now directly uses the keyword
`_Static_assert`.