Add new structure for test results
There is a new enumeration, `CUTE_TestStatus` that defines constants to
represent the exit status of the test: success, failure, etc.
There are two new structures, `CUTE_TestResut`, that stores the name of
a test along with its execution result, and `CUTE_TestCaseOutcome`,
that stores the results of all the tests from a test case.
Since the latter contains a flexible array member, two functions are
also declared to instantiate and deallocate one instance of this
structure.
These declarations are written in a new header, `cute_result.h`, and
the functions are defined in a new source, `cute_result.c`.

The inclusion of this header is added in `cute_test.h`.