Signals handling prototype
Warning: this version is **absolutely not safe**. It uses unsafe
practices that can lead to undefined behaviour and program unreliablity.

This commit introduces the handling of signals emitted by the execution
of the code (e.g. `SIGABRT` or `SIGFPE`) or by the user (`SIGINT`,
`SIGQUIT` or `SIGTSTP`; this category of signals is less concerned
here).
The signals are caught through a function handler set with `sigaction`,
the signal handler calls `longjmp` to return to the running of the
tests (where `setjmp` was called`).

The reason why `longjmp` is used is that it is sort-of a program
interruption. It allows to break the signal transmission and hence the
program can recover from an `abort()` call. However, this causes
undefined behaviour and shall be removed as soon as possible.

For now, the signal handler only prints a message indicating which
signal was caught; it jumps back with a status value that is not even
used -- total WIP!

The macro `_POSIX_C_SOURCE` is defined in the preprocessor flags in the
Makefile to make available the definition of the structure `sigaction`.