Replaced abort() with raise(SIGABRT)
The standard function `abort()` resends the signal after the handler
returns, so in order to recover from it `longjmp()` must be used.
Since we're trying not to, the softer `raise(SIGABRT)` is used instead,
in both the definition of `CUTE_fail` and the abort test. This
instruction has the signal sent just once and can be nicely handled.

The function `CUTE_fail` had to be declared without the *noreturn*
attribute, since `raise` does not necessarily *not return*.