Remove calls to longjmp, simplified test statutes
The `longjmp` calls are removed, since they could produce undefined
behavior.
The variable used to store the test status is made atomic to also avoid
undefined behavior, and the signals handler now only sets this variable
to a value from the enumeration `status`.
This value is used after the signal handling, back in the function
`CUTE_runTestCase`, where an appropriate (but terse) message is printed
on `stderr`.