Add new status for a test: STATUS_OK
The status' value is `0`, it means that the test succeeded.
It is the initial value for the shared variable `_status`, set before
running each test. It is as before set to an error value in the signal
handler, so no signal sent means success.
In case the test is successful (ie. `_status` is still equal to
`STATUS_OK` after the test is run), a debug message is output, printing
only `success`.