Fix function CUTE_runTestCase
The function has been rearranged so that test status is actually taken
into acount: if the test passed, the number of successes in the case
outcome is incremented; if a cancellation signal is caught, all the
remaining tests in the case are cancelled.
The before and after functions are called respectively just before and
after the test procedure itself. An empty `case` clause is added to the
`switch` statement that handles the value of the test status for the
constant `CUTE_RESULT_IGNORED` to shut up a warning: