Tests whose name start with I are ignored
The function `CUTE_runTestCase` takes into account the name of each
test: if the first character is an uppercase `I`, the test is not
run and its status is set to `CUTE_RESULT_IGNORED`. The name of the
test in the result is set to the name of the test + 1, to remove the
leading `'I'`.

A test of the feature is added to the test suites.