Add a title to the test case
The title allows to pass information to the case about the object being
tested by the the case. This increases the relation between the tests
in the case.
The title is then passed to the results of the case when run.

A member called `title` is added to the internal structure for
`CUTE_TestCase` and to the structure `CUTE_RunResults`; an eponymous
parameter is added to the constructor-like functions `CUTE_newTestCase`
and `CUTE_prepareResults` to set the member.
A function `CUTE_getCaseTitle` is added to retrieve this value.

Default titles are given to the test cases in the test source.