Remove access to structure CUTE_TestSuite
The structure is thought to have only one instance at a time; access to
it is removed to enforce this design. A unique internal instance is
declared and the functions operate on that instance only.

The functions are slightly renamed and their pointer argument is
removed, as is the declaration of the structure `CUTE_TestSuite`.

The structure of that instance is also modified to contain the array of
results so that it can be freed in `CUTE_cleanUpTestSuite` along with
the test cases. The flexible array member therefore had to be changed
to an array pointer, like the results array.
These two members are allocated with `calloc(3)` after the instance
itself.