Add declaration of a results logger.
A new header is added that declares an enum (`CUTE_LogFormat`)
representing the logging format, a set of functions to specify a format
and a file to log the results (`CUTE_setLogFormat` and `CUTE_setLogFile`
respectively).
The last function, `CUTE_logResults`, performs the actual logging.

The header is not accessible for direct `#include` directives, but is
provided with `cuite_test.h`.