Remove functions to set log file & format
The removed functions were `CUTE_setLogFile` and `CUTE_setLogFormat`.
They used to set global variables that are replaced with new parameters
to `CUTE_logResults`.
A new function declared inline, `CUTE_printResults` is a helper
function to call `CUTE_logResults` with log file and format parameters
defaulting to `stdout` and simple text format. This is the function
called from the test source.

The logging function is `typedef`ed in the source to factorize some
lines of code.