Fix usage of the Docker image
Sections `before_install`, `install` and `script` of the Travis-CI
config file are changed: the image is now downloaded and run in a
detached container (`before_install`), the repository is downloaded
into the container (`install`) and the build and test commands are run
inside the container (`script`).

Sections are also separated, comments are added for each one.