Run the build and tests from a subshell
In the image, `cd` is a shell built-in and the parser can't find the
executable. The command has thus to be run as an argument to a shell,
here `/bin/sh`.
However, since this command is run in a sub-shell, the directory change
is lost after the subshell exits; the build and tests run must then
execute in the same subshell.