Define new function CUTE_exit()
The function is a wrapper around the standard `exit(3)` function that
sends the proper exit status: `EXIT_SUCCESS` iff all run tests passed,
`EXIT_FAILURE` otherwise (including ignored tests).

The function is declared in `cute_test.h`, however ti is defined in
`cute_suite.c` because it uses a field of the internal suite instance.

A new function attribute is declared, `CUTE_NORETURN`, to indicate that
the function never returns control back to the caller.